Set up player info boxes, add replay feature
Skeleton of the player information boxes has been implemented, for white and black. This will allow things to be added in the future such as time, captured opponent pieces. Also allow the players to play again seamlessly by resetting the game and board (all that needs to be done is that the current game loop is exited and the start function is called again.