Basic timeout vs insuffient material handling
If player A runs out of time, then if Player B only has a lone king, king + knight, or king + bishop, then the game is drawn. Can be improved by algorithm to search for mates in 1-3 or so even with only one knight/bishop (not necessarily forced, but possible), better than allowing actually winning situations for player B to not be noticed. Lone king cannot win, rook/pawn/queen always can even if not forcefully possible.